--- javascript/javascript3/week1/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Quiz App</title>
-  </head>
-
-  <body>
-    <p>Redirecting to <a href="quizzForm.html">Create 4 choice Quiz</a>...</p>
-
-    <main>
-      <section id="game-mode">
-        <h2 class="game-mode-title">Game Mode</h2>
-        <div class="mode-container">
-          <div class="player">
-            <h3>One Player</h3>
-            <img src="#" alt="oneplayer" />
-            <button id="to-oneplayer" class="to-game">Go</button>
-          </div>
-          <div class="player">
-            <h3>Multiplayer</h3>
-            <img src="#" alt="multiPlayer" />
-            <button id="to-multiplayer" class="to-game">Go</button>
-          </div>
-        </div>
-      </section>
-    </main>
-
-    <script type="module" src="./scripts/script.js">
-      window.location.href = "quizzForm.html";
-    </script>
-  </body>
-</html>
