--- javascript/javascript3/week1/multiPlayer.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>One Player Quiz</title>
-  </head>
-  <body>
-    <main>
-      <section class="quiz-container">
-        <h2 class="player-title">Player 1</h2>
-        <h3 class="quantity-outof">1 out of 10</h3>
-        <div class="point-number-holder">
-          <p class="question-number">Question</p>
-          <p class="point-holder">Point</p>
-        </div>
-        <h2 class="question-title">question</h2>
-        <div class="buttons-holder">
-          <button class="buttons-player1" value="true">True</button>
-          <button class="buttons-player1" value="false">False</button>
-        </div>
-      </section>
-      <section class="quiz-container">
-        <h2 class="player-title">Player 2</h2>
-        <h3 class="quantity-outof">1 out of 10</h3>
-        <div class="point-number-holder">
-          <p class="question-number">Question</p>
-          <p class="point-holder">Point</p>
-        </div>
-        <h2 class="question-title">question</h2>
-        <div class="buttons-holder">
-          <button class="buttons-player2" value="true">True</button>
-          <button class="buttons-player2" value="false">False</button>
-        </div>
-      </section>
-    </main>
-
-    <script type="module" src="./scripts/multiPlayer.js"></script>
-  </body>
-</html>
