--- javascript/javascript3/week1/onePlayer.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>One Player Quiz</title>
-  </head>
-  <body>
-    <main>
-      <section id="quiz-container">
-        <h3 id="quantity-outof"></h3>
-        <div id="point-number-holder">
-          <p id="question-number"></p>
-          <p id="point-holder"></p>
-        </div>
-        <h2 id="question-title"></h2>
-        <div id="buttons-holder">
-          <button id="true-button" value="true">True</button>
-          <button id="false-button" value="false">False</button>
-        </div>
-      </section>
-    </main>
-
-    <script src="./scripts/onePlayer.js" type="module"></script>
-  </body>
-</html>
