--- javascript/javascript3/week1/quizForUser.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./styles/quizForUser.css">
-    <title>Quiz For you</title>
-</head>
-<body>
-    <script type="module" src="./scripts/quizForYou.js"></script>
-</body>
-</html>
