--- javascript/javascript3/week1/readme.md
@@ -1,1 +1,0 @@
-# Readme
