--- javascript/javascript3/week1/scripts/questions.js
@@ -1,52 +1,0 @@
-
-export const questions = [
-  {
-      question: "The Great Wall of China is visible from space.",
-      answer: false
-    },
-    {
-      question: "The capital of Australia is Sydney.",
-      answer: false
-    },
-    
-    // Science
-    {
-      question: "Water boils at 100 degrees Celsius at sea level.",
-      answer: true
-    },
-    {
-      question: "Humans share about 98% of their DNA with chimpanzees.",
-      answer: true
-    },
-    
-    // History
-    {
-      question: "The Titanic sank on its maiden voyage in 1912.",
-      answer: true
-    },
-    {
-      question: "The Roman Empire fell in the 5th century AD.",
-      answer: true
-    },
-    
-    // Geography
-    {
-      question: "Africa is the largest continent in the world by land area.",
-      answer: false
-    },
-    {
-      question: "Mount Everest is located in the Andes Mountains.",
-      answer: false
-    },
-    
-    // Pop Culture
-    {
-      question: "The first Harry Potter book was published in 1997.",
-      answer: true
-    },
-    {
-      question: "The Beatles were originally from Ireland.",
-      answer: false
-    }
-  ];
-
