--- javascript/javascript3/week1/scripts/script.js
@@ -1,9 +1,0 @@
-const buttonToOne = document.getElementById("to-oneplayer");
-buttonToOne.addEventListener("click", () => {
-  window.location.href = "onePlayer.html";
-});
-
-const buttonToMulti = document.querySelector("#to-multiplayer");
-buttonToMulti.addEventListener("click", () => {
-  window.location.href = "multiPlayer.html";
-});
