+++ javascript/javascript3/week2/quizitems.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./styles/styles.css">
+    <title>Quiz Generator</title>
+</head>
+<body>
+<script type="module" src="./scripts/quizItems.js"></script>
+</body>
+</html>
+  <!-- <form id="formPost">
+        <label for="question">Write your question.</label><br>
+        <input id="question" name="question" type="text" width="170" required/>
+
+        <p>Options:</p>
+        <ul>
+           <li><label for="option-one">1st</label></li>
+           <input id="option-one" name="option-one" type="text" required/> <br>
+           <li><label for="option-two">2nd</label></li>
+           <input id="option-two" name="option-two" type="text" required/> <br>
+           <li><label for="option-three">3rd</label></li>
+           <input id="option-three" name="option-three" type="text" required/> <br>
+           <li><label for="option-four">4th</label></li>
+           <input id="option-four" name="option-four" type="text" required/> <br>
+        </ul>
+        <h4>Tick the correct answer:</h4>
+        <div id="radio-items">
+            <label>1st</label>
+            <input class='question-one-options' type="radio" name="isCorrect" required/>
+            <label>2nd</label>
+            <input class='question-one-options' type="radio" name="isCorrect" required/>
+            <label>3rd</label>
+            <input class='question-one-options' type="radio" name="isCorrect" required/>
+            <label>4th</label>
+            <input class='question-one-options' type="radio" name="isCorrect" required/>
+        </div>
+        <button type="submit" id="saveButton">Save</button>
+    </form>
+     -->
