+++ javascript/javascript3/week2/quizzForm.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+
+
+    
+  
+<script type="module" src="./scripts/quzzform.js"></script>
+
+</body>
+</html>
