+++ javascript/javascript3/week2/scripts/script.js
@@ -1,0 +1,9 @@
+const buttonToOne = document.getElementById("to-oneplayer");
+buttonToOne.addEventListener("click", () => {
+  window.location.href = "onePlayer.html";
+});
+
+const buttonToMulti = document.querySelector("#to-multiplayer");
+buttonToMulti.addEventListener("click", () => {
+  window.location.href = "multiPlayer.html";
+});
