+++ javascript/javascript3/week2/scripts/token.js
