+++ javascript/javascript3/week2/styles/quizForUser.css
@@ -1,0 +1,76 @@
+body{
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    height: 100%;
+    
+}
+
+#form-container{
+    width: 500px;
+    max-width: 1300px;
+
+
+    border-top: 1px solid black;
+    border-bottom: 1px solid black;
+    margin: 20px auto;
+
+}
+form{
+    width: 100%;
+    margin: 20px auto;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    flex-wrap: wrap;
+    border-radius: 30px;
+    background-color: antiquewhite;
+    padding: 10px 0;
+}
+
+form label{
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    text-align: center;
+    margin: 20px 0;
+
+}
+form div{
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    align-items: center;
+    justify-content: center;
+}
+form div label{
+    border-bottom: 2px solid grey;
+    padding: 10px;
+    margin: 0 10px;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    font-size: 13px;
+    position:relative;
+}
+form div label input{
+   width: 15px;
+   height: 15px;
+   position:absolute;
+   top: 45px;
+   left: 16px;
+   
+
+}
+form button{
+    padding: 5px 20px;
+    margin: 0 0 20px 0;
+}
+.quizLabelID{
+    font-size: 10px;
+    font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+}
+.questionNumber{
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    font-size: 13px;
+    text-align: left;
+    width: 80%;
+}
