+++ javascript/javascript3/week2/styles/styles.css
@@ -1,0 +1,128 @@
+body{
+    width: 100%;
+    height: 100%;
+    box-sizing: border-box;
+    margin: 0;
+    
+}
+form{
+    background-color: rgba(213, 241, 180, 0.5);
+    margin: 20px auto;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    place-items: center;
+    width: 500px;
+    padding: 30px;
+    border-radius: 30px;
+}
+
+.questionLabel{
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    padding: 10px 20px;
+}
+
+.questionInput{
+    font-style: normal;
+    border: 2px solid grey;
+    width: 100%;
+    height: 30px;
+    border-radius: 5px;
+    box-shadow: 0;
+    text-shadow: 0;
+    margin: 20px;
+}
+.optContainer{
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+    place-items: center;
+    
+}
+.optLabel{
+    width: 80%;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    font-size: 14px;
+    text-align: center;
+    margin: 20px 0 5px 0;
+
+
+}
+.optContainer input{
+    font-style: normal;
+    border: 2px solid grey;
+    width: 40%;
+    height: 30px;
+    border-radius: 5px;
+    box-shadow: 0;
+    text-shadow: 0;
+
+}
+
+.ulOptions{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    justify-items: center;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    background-color: rgba(71, 71, 71, 0.1);
+    font-size: 12px;
+    border-radius: 20px;
+    margin-bottom: 20px;
+    
+}
+.eachOption{
+    display:flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    justify-items: center;
+    padding: 15px 15px 0 15px;
+    margin: 10px;
+    border: 10px solid grey;
+    border-radius: 10px;
+    border-right: rgba(255, 255, 255, 0.5);
+    border-bottom: rgba(255, 255, 255, 0.5);
+    border-left: rgba(0, 0, 0,  0.1);
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+}
+.eachOption label{
+    margin-bottom: 12px;
+}
+.eachOption input{
+    width: 20px;
+    height: 20px;
+    border: 2px solid grey;
+    padding-bottom: 0;
+    margin-bottom: 0;
+}
+button{
+    padding: 10px 40px;
+    border: 0;
+    border-radius: 10px;
+    background-color:  rgb(255, 255, 255);
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+}
+.quizContainer{
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+}
+.questionTitle{
+    width: 100%;
+    text-align: center;
+}
+.optionsitem{
+    margin: 20px;
+    padding: 10px;
+    font-weight: 400;
+    font-size: 14px;
+}
